EFS-435 Add Sentry integration
