+++ package.json
@@ -117,6 +117,7 @@
     "@mongodb-js/zstd": "^1.1.0",
     "@opensearch-project/opensearch": "^2.2.0",
     "@popperjs/core": "^2.11.8",
+    "@sentry/node": "^6.19.7",
     "accepts": "^1.3.8",
     "async": "^3.2.4",
     "async-mutex": "^0.4.0",
