+++ src/dataLayer/databaseUpdateManager.js
@@ -8,7 +8,7 @@ const {ResourceLocatorFactory} = require('../operations/common/resourceLocatorFa
 const {RethrownError} = require('../utils/rethrownError');
 const {ResourceMerger} = require('../operations/common/resourceMerger');
 const {PreSaveManager} = require('../preSaveHandlers/preSave');
-const {logTraceSystemEventAsync} = require('../operations/common/logging');
+const {logTraceSystemEventAsync} = require('../operations/common/systemEventLogging');
 const {DatabaseQueryFactory} = require('./databaseQueryFactory');
 const {ConfigManager} = require('../utils/configManager');
 const {getCircularReplacer} = require('../utils/getCircularReplacer');
