+++ src/utils/changeEventProducer.js
@@ -3,7 +3,7 @@ const {generateUUID} = require('./uid.util');
 const moment = require('moment-timezone');
 const {assertTypeEquals, assertIsValid} = require('./assertType');
 const {ResourceManager} = require('../operations/common/resourceManager');
-const {logTraceSystemEventAsync} = require('../operations/common/logging');
+const {logTraceSystemEventAsync} = require('../operations/common/systemEventLogging');
 const AuditEvent = require('../fhir/classes/4_0_0/resources/auditEvent');
 const CodeableConcept = require('../fhir/classes/4_0_0/complex_types/codeableConcept');
 const Coding = require('../fhir/classes/4_0_0/complex_types/coding');
